backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
httpx==0.27.0
Pillow==10.4.0
pydantic==2.9.0
volcengine
tos
# --- 数据库 + 认证（Phase 1 新增） ---
sqlalchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
python-dotenv>=1.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3